requirements.txt
# Core libraries for Reddit automation dashboard
streamlit>=1.32.0
praw>=7.7.1
APScheduler>=3.10.4
pandas>=2.0.0

# AI/ML libraries - Groq API integration
groq>=0.4.1
openai>=1.3.0

# Database and scheduling
sqlalchemy>=2.0.0

# Additional utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# For data analysis and visualization
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.0.0
